feat: deleteNote/s and edit
